+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Fixed
 - Fixed issue where `UseStrategy`, `ExtendsParseStrategy` and `ImplementsParseStrategy` resulted in partial symbol names
 ### Added
+- Added `TypedAttributeStrategy` to recognize usage for consumed full qualified property types
 ### Changed
 ### Removed